requirements.txt
PyYAML~=5.4.1
cryptography~=2.9.2
snowflake-connector-python~=2.3.10